pyproject.toml
[tool.poetry]
name = "gdelt_ml_pipeline"
version = "0.1.0"
description = "ML Pipeline for GDELT"
authors = ["Raffay Rana <ar534@duke.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pytorch-forecasting = "^1.1.1"
torch = "^2.4.1"
transformers = "^4.45.1"
yfinance = "^0.2.44"
ta = "^0.11.0"
textblob = "^0.18.0.post0"
pytorch-lightning = "^2.4.0"
optuna = "^4.0.0"
shap = "^0.46.0"
pandas = "^2.2.3"
numpy = "<2.0.0"
matplotlib = "^3.9.2"
google-cloud-bigquery = "^3.26.0"
openai = "^1.51.0"
tenacity = "^9.0.0"
python-dotenv = "^1.0.1"
optuna-integration = {extras = ["pytorch-lightning"], version = "^4.0.0"}
db-dtypes = "^1.3.0"
google-cloud-bigquery-storage = "^2.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
